src/adapter/requirements.txt
cftime==1.0.4.2
netCDF4==1.5.3
numpy==1.18.1
pandas==0.25.3
progressbar2==3.47.0
python-dateutil==2.8.1
python-utils==2.3.0
pytz==2019.3
scipy==1.4.1
six==1.13.0
xarray==0.14.1
